started Model log_prob function --- beginning elbo deriv re-factorization
