+++ src/ElboDeriv.jl
@@ -4,46 +4,18 @@ Calculate value, gradient, and hessian of the variational ELBO.
 module ElboDeriv
 
 using ..Model
-using ..SensitiveFloats
+#using ..SensitiveFloats
 import ..WCSUtils
 import ..SensitiveFloats.clear!
 
-export ElboArgs
 
-
-"""
-ElboArgs stores the arguments needed to evaluate the variational objective
-function
-"""
-type ElboArgs{NumType <: Number}
-    S::Int64
-    N::Int64
-    images::Vector{TiledImage}
-    vp::VariationalParams{NumType}
-    tile_source_map::Vector{Matrix{Vector{Int}}}
-    patches::Matrix{SkyPatch}
-    active_sources::Vector{Int}
-end
-
-
-function ElboArgs{NumType <: Number}(
-            images::Vector{TiledImage},
-            vp::VariationalParams{NumType},
-            tile_source_map::Vector{Matrix{Vector{Int}}},
-            patches::Matrix{SkyPatch},
-            active_sources::Vector{Int})
-    N = length(images)
-    S = length(vp)
-    @assert length(tile_source_map) == N
-    @assert size(patches, 1) == S
-    @assert size(patches, 2) == N
-    ElboArgs(S, N, images, vp, tile_source_map, patches, active_sources)
-end
-
+# TODO make the functions dependent on these objects dependent on ..Model instead
+import ..Model: ElboArgs, ActivePixel #, SensitiveFloats
+export ElboArgs, ActivePixel #, SensitiveFloats
 
 include("elbo_kl.jl")
 include("source_brightness.jl")
-include("bivariate_normals.jl")
+#include("bivariate_normals.jl")
 
 
 """
@@ -117,21 +89,6 @@ function load_bvn_mixtures{NumType <: Number}(
 end
 
 
-
-# TODO: the identification of active pixels should go in pre-processing
-type ActivePixel
-    # image index
-    n::Int
-
-    # Linear tile index:
-    tile_ind::Int
-
-    # Location in tile:
-    h::Int
-    w::Int
-end
-
-
 """
 Store pre-allocated memory in this data structures, which contains
 intermediate values used in the ELBO calculation.
@@ -157,105 +114,6 @@ function HessianSubmatrices(NumType::DataType, i::Int)
     u_u = zeros(NumType, 2, 2)
     shape_shape = zeros(NumType, shape_p, shape_p)
     HessianSubmatrices{NumType}(u_u, shape_shape)
-end
-
-
-type ElboIntermediateVariables{NumType <: Number}
-
-    bvn_derivs::BivariateNormalDerivatives{NumType}
-
-    # Vectors of star and galaxy bvn quantities from all sources for a pixel.
-    # The vector has one element for each active source, in the same order
-    # as ea.active_sources.
-
-    # TODO: you can treat this the same way as E_G_s and not keep a vector around.
-    fs0m_vec::Vector{SensitiveFloat{StarPosParams, NumType}}
-    fs1m_vec::Vector{SensitiveFloat{GalaxyPosParams, NumType}}
-
-    # Brightness values for a single source
-    E_G_s::SensitiveFloat{CanonicalParams, NumType}
-    E_G2_s::SensitiveFloat{CanonicalParams, NumType}
-    var_G_s::SensitiveFloat{CanonicalParams, NumType}
-
-    # Subsets of the Hessian of E_G_s and E_G2_s that allow us to use BLAS
-    # functions to accumulate Hessian terms. There is one submatrix for
-    # each celestial object type in 1:Ia
-    E_G_s_hsub_vec::Vector{HessianSubmatrices{NumType}}
-    E_G2_s_hsub_vec::Vector{HessianSubmatrices{NumType}}
-
-    # Expected pixel intensity and variance for a pixel from all sources.
-    E_G::SensitiveFloat{CanonicalParams, NumType}
-    var_G::SensitiveFloat{CanonicalParams, NumType}
-
-    # Pre-allocated memory for the gradient and Hessian of combine functions.
-    combine_grad::Vector{NumType}
-    combine_hess::Matrix{NumType}
-
-    # A placeholder for the log term in the ELBO.
-    elbo_log_term::SensitiveFloat{CanonicalParams, NumType}
-
-    # The ELBO itself.
-    elbo::SensitiveFloat{CanonicalParams, NumType}
-
-    # If false, do not calculate hessians or derivatives.
-    calculate_derivs::Bool
-
-    # If false, do not calculate hessians.
-    calculate_hessian::Bool
-end
-
-
-"""
-Args:
-    - S: The total number of sources
-    - num_active_sources: The number of actives sources (with deriviatives)
-    - calculate_derivs: If false, only calculate values
-    - calculate_hessian: If false, only calculate gradients. Note that if
-                calculate_derivs = false, then hessians will not be
-                calculated irrespective of the value of calculate_hessian.
-"""
-function ElboIntermediateVariables(NumType::DataType,
-                                   S::Int,
-                                   num_active_sources::Int;
-                                   calculate_derivs::Bool=true,
-                                   calculate_hessian::Bool=true)
-    @assert NumType <: Number
-
-    bvn_derivs = BivariateNormalDerivatives{NumType}(NumType)
-
-    # fs0m and fs1m accumulate contributions from all bvn components
-    # for a given source.
-    fs0m_vec = Array(SensitiveFloat{StarPosParams, NumType}, S)
-    fs1m_vec = Array(SensitiveFloat{GalaxyPosParams, NumType}, S)
-    for s = 1:S
-        fs0m_vec[s] = zero_sensitive_float(StarPosParams, NumType)
-        fs1m_vec[s] = zero_sensitive_float(GalaxyPosParams, NumType)
-    end
-
-    E_G_s = zero_sensitive_float(CanonicalParams, NumType, 1)
-    E_G2_s = zero_sensitive_float(CanonicalParams, NumType, 1)
-    var_G_s = zero_sensitive_float(CanonicalParams, NumType, 1)
-
-    E_G_s_hsub_vec =
-        HessianSubmatrices{NumType}[ HessianSubmatrices(NumType, i) for i=1:Ia ]
-    E_G2_s_hsub_vec =
-        HessianSubmatrices{NumType}[ HessianSubmatrices(NumType, i) for i=1:Ia ]
-
-    E_G = zero_sensitive_float(CanonicalParams, NumType, num_active_sources)
-    var_G = zero_sensitive_float(CanonicalParams, NumType, num_active_sources)
-
-    combine_grad = zeros(NumType, 2)
-    combine_hess = zeros(NumType, 2, 2)
-
-    elbo_log_term =
-        zero_sensitive_float(CanonicalParams, NumType, num_active_sources)
-    elbo = zero_sensitive_float(CanonicalParams, NumType, num_active_sources)
-
-    ElboIntermediateVariables{NumType}(
-        bvn_derivs, fs0m_vec, fs1m_vec,
-        E_G_s, E_G2_s, var_G_s, E_G_s_hsub_vec, E_G2_s_hsub_vec,
-        E_G, var_G, combine_grad, combine_hess,
-        elbo_log_term, elbo, calculate_derivs, calculate_hessian)
 end
 
 
@@ -417,61 +275,6 @@ function accum_galaxy_pos!{NumType <: Number}(
     end # if calculate_derivs
 end
 
-
-"""
-Populate fs0m_vec and fs1m_vec for all sources for a given pixel.
-
-Args:
-    - elbo_vars: Elbo intermediate values.
-    - ea: Model parameters
-    - tile_sources: A vector of integers of sources in 1:ea.S affecting the tile
-    - tile: An ImageTile
-    - h, w: The integer locations of the pixel within the tile
-    - gal_mcs: Galaxy components
-    - star_mcs: Star components
-
-Returns:
-    Updates elbo_vars.fs0m_vec and elbo_vars.fs1m_vec in place with the total
-    shape contributions to this pixel's brightness.
-"""
-function populate_fsm_vecs!{NumType <: Number}(
-                    elbo_vars::ElboIntermediateVariables{NumType},
-                    ea::ElboArgs{NumType},
-                    tile_sources::Vector{Int},
-                    tile::ImageTile,
-                    h::Int, w::Int,
-                    gal_mcs::Array{GalaxyCacheComponent{NumType}, 4},
-                    star_mcs::Array{BvnComponent{NumType}, 2})
-    x = Float64[tile.h_range[h], tile.w_range[w]]
-    for s in tile_sources
-        # ensure tile.b is a filter band, not an image's index
-        @assert 1 <= tile.b <= B
-        wcs_jacobian = ea.patches[s, tile.b].wcs_jacobian
-        active_source = s in ea.active_sources
-
-        calculate_hessian =
-            elbo_vars.calculate_hessian && elbo_vars.calculate_derivs && active_source
-        clear!(elbo_vars.fs0m_vec[s], calculate_hessian)
-        for k = 1:psf_K # PSF component
-            accum_star_pos!(
-                elbo_vars, s, star_mcs[k, s], x, wcs_jacobian, active_source)
-        end
-
-        clear!(elbo_vars.fs1m_vec[s], calculate_hessian)
-        for i = 1:2 # Galaxy types
-            for j in 1:8 # Galaxy component
-                # If i == 2 then there are only six galaxy components.
-                if (i == 1) || (j <= 6)
-                    for k = 1:psf_K # PSF component
-                        accum_galaxy_pos!(
-                            elbo_vars, s, gal_mcs[k, j, i, s], x, wcs_jacobian,
-                            active_source)
-                    end
-                end
-            end
-        end
-    end
-end
 
 
 """
@@ -981,32 +784,6 @@ function tile_predicted_image{NumType <: Number}(
     tile_predicted_image(
         elbo_vars, tile, ea, tile_sources, sbs, star_mcs, gal_mcs,
         include_epsilon=include_epsilon)
-end
-
-
-"""
-Get the active pixels (pixels for which the active sources are present).
-TODO: move this to pre-processing and use it instead of setting low-signal
-pixels to NaN.
-"""
-function get_active_pixels{NumType <: Number}(
-                    ea::ElboArgs{NumType})
-    active_pixels = ActivePixel[]
-
-    for n in 1:ea.N, tile_ind in 1:length(ea.images[n].tiles)
-        tile_sources = ea.tile_source_map[n][tile_ind]
-        if length(intersect(tile_sources, ea.active_sources)) > 0
-            tile = ea.images[n].tiles[tile_ind]
-            h_width, w_width = size(tile.pixels)
-            for w in 1:w_width, h in 1:h_width
-                if !Base.isnan(tile.pixels[h, w])
-                    push!(active_pixels, ActivePixel(n, tile_ind, h, w))
-                end
-            end
-        end
-    end
-
-    active_pixels
 end
