+++ test/test_log_prob.jl
@@ -1,0 +1,11 @@
+## test the main entry point in Celeste: the `infer` function
+#import JLD
+
+"""
+test log_prob with multiple settings of parameters --- make sure ll values
+differ
+"""
+function test_log_prob_values()
+end
+
+test_log_prob_values()
